+++ src/ui/footer/components/InteractivePiano/InteractivePiano.h
@@ -2,36 +2,20 @@
 
 #include <JuceHeader.h>
 #include "ui/extra/Component/Panel/ColoredPanel.h"
-#include "utils/FontManager.h"
 #include "InteractivePianoContentArea.h"
+
+// Forward declarations
+class FontManager;
 
 //==============================================================================
 // Conteneur principal avec padding
 class InteractivePiano : public ColoredPanel
 {
 public:
-    InteractivePiano(juce::Colour color) 
-        : ColoredPanel(color),
-          fontManager(),
-          contentArea()
-    {
-        addAndMakeVisible(contentArea);
-    }
+    InteractivePiano(juce::Colour color);
     
-    void paint(juce::Graphics& g) override
-    {
-        // Appeler le paint de ColoredPanel pour le fond
-        ColoredPanel::paint(g);
-    }
-    
-    void resized() override
-    {
-        // Appliquer un padding de 10px tout autour
-        auto bounds = getLocalBounds();
-        auto contentBounds = bounds.reduced(10);
-        
-        contentArea.setBounds(contentBounds);
-    }
+    void paint(juce::Graphics& g) override;
+    void resized() override;
     
 private:
     juce::SharedResourcePointer<FontManager> fontManager;
